Starting rename of executable_layout -> pipeline_layout.
Progress on #10144.
