+++ experimental/rocm/CMakeLists.txt
@@ -35,14 +35,14 @@ iree_cc_library(
     "rocm_event.h"
     "event_semaphore.c"
     "event_semaphore.h"
-    "executable_layout.c"
-    "executable_layout.h"
     "direct_command_buffer.c"
     "direct_command_buffer.h"
     "native_executable.c"
     "native_executable.h"
     "nop_executable_cache.c"
     "nop_executable_cache.h"
+    "pipeline_layout.c"
+    "pipeline_layout.h"
     "status_util.c"
     "status_util.h"
   INCLUDES
