+++ experimental/rocm/native_executable.c
@@ -9,7 +9,7 @@
 #include <stddef.h>
 
 #include "experimental/rocm/dynamic_symbols.h"
-#include "experimental/rocm/executable_layout.h"
+#include "experimental/rocm/pipeline_layout.h"
 #include "experimental/rocm/status_util.h"
 #include "iree/base/api.h"
 #include "iree/base/tracing.h"
