+++ experimental/rocm/rocm_device.c
@@ -14,8 +14,8 @@
 #include "experimental/rocm/direct_command_buffer.h"
 #include "experimental/rocm/dynamic_symbols.h"
 #include "experimental/rocm/event_semaphore.h"
-#include "experimental/rocm/executable_layout.h"
 #include "experimental/rocm/nop_executable_cache.h"
+#include "experimental/rocm/pipeline_layout.h"
 #include "experimental/rocm/rocm_allocator.h"
 #include "experimental/rocm/rocm_event.h"
 #include "experimental/rocm/status_util.h"
