+++ runtime/iree.natvis
@@ -459,7 +459,7 @@
 <!--=========================================================================-->
 
 <!--=========================================================================-->
-<!-- iree/hal/executable_layout.h -->
+<!-- iree/hal/pipeline_layout.h -->
 <!--=========================================================================-->
 
 <!--=========================================================================-->
