+++ runtime/src/iree/hal/BUILD
@@ -51,10 +51,10 @@ iree_runtime_cc_library(
         "executable.h",
         "executable_cache.c",
         "executable_cache.h",
-        "executable_layout.c",
-        "executable_layout.h",
         "fence.c",
         "fence.h",
+        "pipeline_layout.c",
+        "pipeline_layout.h",
         "resource.h",
         "semaphore.c",
         "semaphore.h",
