+++ runtime/src/iree/hal/CMakeLists.txt
@@ -44,10 +44,10 @@ iree_cc_library(
     "executable.h"
     "executable_cache.c"
     "executable_cache.h"
-    "executable_layout.c"
-    "executable_layout.h"
     "fence.c"
     "fence.h"
+    "pipeline_layout.c"
+    "pipeline_layout.h"
     "resource.h"
     "semaphore.c"
     "semaphore.h"
