+++ runtime/src/iree/hal/api.h
@@ -9,21 +9,21 @@
 #ifndef IREE_HAL_API_H_
 #define IREE_HAL_API_H_
 
-#include "iree/hal/allocator.h"          // IWYU pragma: export
-#include "iree/hal/buffer.h"             // IWYU pragma: export
-#include "iree/hal/buffer_view.h"        // IWYU pragma: export
-#include "iree/hal/buffer_view_util.h"   // IWYU pragma: export
-#include "iree/hal/command_buffer.h"     // IWYU pragma: export
-#include "iree/hal/device.h"             // IWYU pragma: export
-#include "iree/hal/driver.h"             // IWYU pragma: export
-#include "iree/hal/driver_registry.h"    // IWYU pragma: export
-#include "iree/hal/event.h"              // IWYU pragma: export
-#include "iree/hal/executable.h"         // IWYU pragma: export
-#include "iree/hal/executable_cache.h"   // IWYU pragma: export
-#include "iree/hal/executable_layout.h"  // IWYU pragma: export
-#include "iree/hal/fence.h"              // IWYU pragma: export
-#include "iree/hal/resource.h"           // IWYU pragma: export
-#include "iree/hal/semaphore.h"          // IWYU pragma: export
-#include "iree/hal/string_util.h"        // IWYU pragma: export
+#include "iree/hal/allocator.h"         // IWYU pragma: export
+#include "iree/hal/buffer.h"            // IWYU pragma: export
+#include "iree/hal/buffer_view.h"       // IWYU pragma: export
+#include "iree/hal/buffer_view_util.h"  // IWYU pragma: export
+#include "iree/hal/command_buffer.h"    // IWYU pragma: export
+#include "iree/hal/device.h"            // IWYU pragma: export
+#include "iree/hal/driver.h"            // IWYU pragma: export
+#include "iree/hal/driver_registry.h"   // IWYU pragma: export
+#include "iree/hal/event.h"             // IWYU pragma: export
+#include "iree/hal/executable.h"        // IWYU pragma: export
+#include "iree/hal/executable_cache.h"  // IWYU pragma: export
+#include "iree/hal/fence.h"             // IWYU pragma: export
+#include "iree/hal/pipeline_layout.h"   // IWYU pragma: export
+#include "iree/hal/resource.h"          // IWYU pragma: export
+#include "iree/hal/semaphore.h"         // IWYU pragma: export
+#include "iree/hal/string_util.h"       // IWYU pragma: export
 
 #endif  // IREE_HAL_API_H_
