+++ runtime/src/iree/hal/command_buffer.h
@@ -15,7 +15,7 @@
 #include "iree/hal/buffer.h"
 #include "iree/hal/event.h"
 #include "iree/hal/executable.h"
-#include "iree/hal/executable_layout.h"
+#include "iree/hal/pipeline_layout.h"
 #include "iree/hal/resource.h"
 
 #ifdef __cplusplus
