+++ runtime/src/iree/hal/command_buffer_validation.c
@@ -18,7 +18,7 @@
 #include "iree/hal/device.h"
 #include "iree/hal/event.h"
 #include "iree/hal/executable.h"
-#include "iree/hal/executable_layout.h"
+#include "iree/hal/pipeline_layout.h"
 #include "iree/hal/resource.h"
 
 // Returns success iff the queue supports the given command categories.
