+++ runtime/src/iree/hal/cts/pipeline_layout_test.h
@@ -4,8 +4,8 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef IREE_HAL_CTS_EXECUTABLE_LAYOUT_TEST_H_
-#define IREE_HAL_CTS_EXECUTABLE_LAYOUT_TEST_H_
+#ifndef IREE_HAL_CTS_PIPELINE_LAYOUT_TEST_H_
+#define IREE_HAL_CTS_PIPELINE_LAYOUT_TEST_H_
 
 #include "iree/base/api.h"
 #include "iree/hal/api.h"
@@ -17,9 +17,9 @@ namespace iree {
 namespace hal {
 namespace cts {
 
-class executable_layout_test : public CtsTestBase {};
+class pipeline_layout_test : public CtsTestBase {};
 
-TEST_P(executable_layout_test, CreateWithNoLayouts) {
+TEST_P(pipeline_layout_test, CreateWithNoLayouts) {
   iree_hal_executable_layout_t* executable_layout = NULL;
   IREE_ASSERT_OK(iree_hal_executable_layout_create(
       device_, /*push_constants=*/0, /*set_layout_count=*/0, NULL,
@@ -28,7 +28,7 @@ TEST_P(executable_layout_test, CreateWithNoLayouts) {
   iree_hal_executable_layout_release(executable_layout);
 }
 
-TEST_P(executable_layout_test, CreateWithPushConstants) {
+TEST_P(pipeline_layout_test, CreateWithPushConstants) {
   iree_hal_executable_layout_t* executable_layout = NULL;
   // Note: The Vulkan maxPushConstantsSize limit must be at least 128 bytes:
   // https://www.khronos.org/registry/vulkan/specs/1.2/html/vkspec.html#limits-minmax
@@ -39,7 +39,7 @@ TEST_P(executable_layout_test, CreateWithPushConstants) {
   iree_hal_executable_layout_release(executable_layout);
 }
 
-TEST_P(executable_layout_test, CreateWithOneLayout) {
+TEST_P(pipeline_layout_test, CreateWithOneLayout) {
   iree_hal_descriptor_set_layout_t* descriptor_set_layout = NULL;
   iree_hal_descriptor_set_layout_binding_t descriptor_set_layout_bindings[] = {
       {/*binding=*/0, /*type=*/IREE_HAL_DESCRIPTOR_TYPE_STORAGE_BUFFER},
@@ -59,7 +59,7 @@ TEST_P(executable_layout_test, CreateWithOneLayout) {
   iree_hal_descriptor_set_layout_release(descriptor_set_layout);
 }
 
-TEST_P(executable_layout_test, CreateWithTwoLayouts) {
+TEST_P(pipeline_layout_test, CreateWithTwoLayouts) {
   iree_hal_descriptor_set_layout_t* descriptor_set_layouts[2] = {NULL};
   iree_hal_descriptor_set_layout_binding_t layout_bindings_0[] = {
       {/*binding=*/0, /*type=*/IREE_HAL_DESCRIPTOR_TYPE_STORAGE_BUFFER},
@@ -94,4 +94,4 @@ TEST_P(executable_layout_test, CreateWithTwoLayouts) {
 }  // namespace hal
 }  // namespace iree
 
-#endif  // IREE_HAL_CTS_EXECUTABLE_LAYOUT_TEST_H_
+#endif  // IREE_HAL_CTS_PIPELINE_LAYOUT_TEST_H_
