+++ runtime/src/iree/hal/device.h
@@ -16,7 +16,7 @@
 #include "iree/hal/command_buffer.h"
 #include "iree/hal/event.h"
 #include "iree/hal/executable_cache.h"
-#include "iree/hal/executable_layout.h"
+#include "iree/hal/pipeline_layout.h"
 #include "iree/hal/fence.h"
 #include "iree/hal/resource.h"
 #include "iree/hal/semaphore.h"
