+++ runtime/src/iree/hal/drivers/cuda/CMakeLists.txt
@@ -29,14 +29,14 @@ iree_cc_library(
     "cuda_event.h"
     "event_semaphore.c"
     "event_semaphore.h"
-    "executable_layout.c"
-    "executable_layout.h"
     "graph_command_buffer.c"
     "graph_command_buffer.h"
     "native_executable.c"
     "native_executable.h"
     "nop_executable_cache.c"
     "nop_executable_cache.h"
+    "pipeline_layout.c"
+    "pipeline_layout.h"
     "status_util.c"
     "status_util.h"
     "stream_command_buffer.c"
