+++ runtime/src/iree/hal/drivers/local_task/task_device.c
@@ -17,7 +17,7 @@
 #include "iree/hal/drivers/local_task/task_queue.h"
 #include "iree/hal/drivers/local_task/task_semaphore.h"
 #include "iree/hal/local/local_executable_cache.h"
-#include "iree/hal/local/local_executable_layout.h"
+#include "iree/hal/local/local_pipeline_layout.h"
 #include "iree/hal/utils/buffer_transfer.h"
 
 typedef struct iree_hal_task_device_t {
