+++ runtime/src/iree/hal/drivers/vulkan/BUILD
@@ -40,8 +40,8 @@ iree_runtime_cc_library(
         "native_event.h",
         "native_executable.cc",
         "native_executable.h",
-        "native_executable_layout.cc",
-        "native_executable_layout.h",
+        "native_pipeline_layout.cc",
+        "native_pipeline_layout.h",
         "native_semaphore.cc",
         "native_semaphore.h",
         "nop_executable_cache.cc",
