+++ runtime/src/iree/hal/drivers/vulkan/CMakeLists.txt
@@ -41,8 +41,8 @@ iree_cc_library(
     "native_event.h"
     "native_executable.cc"
     "native_executable.h"
-    "native_executable_layout.cc"
-    "native_executable_layout.h"
+    "native_pipeline_layout.cc"
+    "native_pipeline_layout.h"
     "native_semaphore.cc"
     "native_semaphore.h"
     "nop_executable_cache.cc"
