+++ runtime/src/iree/hal/drivers/vulkan/builtin_executables.cc
@@ -10,7 +10,7 @@
 
 #include "iree/base/tracing.h"
 #include "iree/hal/drivers/vulkan/builtin/builtin_shaders_spv.h"
-#include "iree/hal/drivers/vulkan/native_executable_layout.h"
+#include "iree/hal/drivers/vulkan/native_pipeline_layout.h"
 #include "iree/hal/drivers/vulkan/status_util.h"
 
 namespace iree {
