+++ runtime/src/iree/hal/local/CMakeLists.txt
@@ -96,11 +96,11 @@ iree_cc_library(
     "inline_command_buffer.h"
     "local_executable.h"
     "local_executable_cache.h"
-    "local_executable_layout.h"
+    "local_pipeline_layout.h"
   SRCS
     "inline_command_buffer.c"
     "local_executable_cache.c"
-    "local_executable_layout.c"
+    "local_pipeline_layout.c"
   DEPS
     ::executable_environment
     ::executable_library
