+++ runtime/src/iree/hal/pipeline_layout.c
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/hal/executable_layout.h"
+#include "iree/hal/pipeline_layout.h"
 
 #include <stddef.h>
